fix: stop the monster-HP overflow, persist gathering mastery, repair consumable SQL
Four defects found by a full audit, ordered by blast radius.

Monster milli-HP was stored in an int. Malakor - the authored region 5 boss,
ordinary progression content, not an exploit - ships at 3,000,000 HP, and
3,000,000 * 1000 wraps int to -1,294,967,296. The engine's death check is
CurrentMonsterHp <= 0, so it spawned already dead, paid full kill rewards,
and respawned at the same negative value on the next tick: 6,000,000 XP and
1,500,000 gold per second, indefinitely. 41 of 115 monsters were affected.

This was a regression from the pacing rebalance that set the region bosses
to 3500/14000/82000/440000/3000000; the milli-HP ceiling of 2,147,483 was
never checked against it.

CurrentMonsterHp is now long on both TickStatePayload and
CharacterActivityState, and every spawn site widens before multiplying. The
wire is unchanged - the packet already carried whole HP via a /1000 divide,
so it stays 695 bytes and neither layout guard moves.

The mirror-image defect existed on the damage side: AttackPower * 1000 * 1.5
overflowed for the four strongest monsters, the wrapped negative hit the
Math.Max(1000, ...) floor, and the deadliest monster in the game dealt
exactly 1 HP per swing. Computed in long and saturated.

GetScaledMonsterMaxHp/AttackPower now saturate instead of wrapping. This
matters independently of the data: the endgame multiplier compounds at 1.25
per region tier with no upper bound, so an unchecked cast was guaranteed to
wrap for any player who progressed far enough past the authored tiers.

Gathering mastery had no durable column. Woodcutting and mining mastery were
earned by three separate paths and CONSUMED for gathering yield, but were
absent from both the login hydration and the checkpoint, so every logout
reset both professions to zero - and no UI ever displayed them, so the loss
was invisible. Added four columns, hydration, write-back (safe as an
absolute snapshot here, since the tick thread is the sole author, unlike
TotalItemsCrafted), the two missing proxy mirrors for the levels, and a
panel that shows both professions with progress to the next level.

The ConsumeConsumable handler was the only raw SQL in the codebase with
unquoted identifiers: it queried equipment_instances, which folds to
lowercase and does not exist, and bound BaseItemId as text against an
integer column. Either fault throws, and that lambda's failure handler
force-disconnects the player. Unreachable today only because no button is
bound to the command.

Help texts are pinned to fixed layout heights, so 25 of them rendered
outside their panels. Auto-sizing rather than ellipsis: these are
explanatory strings whose value is being read, and truncating would discard
exactly the part not yet read. Down to 8, all from other creators.
VillageResourceStrip moved to -432 to clear the grown stats panel.

Verified: 188/188 tests green. Reverting CurrentMonsterHp to int now fails
compilation in 10 places, so the widening cannot silently regress; a
reflection assertion pins the field types as a second layer. Scene rebuilt
and saved, then re-audited live: 0 of 35 filled bars lack a sprite, the
mastery panel is present and fully bound.

+++ server/FolkIdle.Server/Engine/CharacterActivityState.cs
@@ -33,7 +33,8 @@ namespace FolkIdle.Server.Engine
 
         // Combat
         public int CurrentMonsterId;
-        public int CurrentMonsterHp;
+        // Milli-HP. Long for the same reason as TickStatePayload's copy.
+        public long CurrentMonsterHp;
         public int PlayerHp;
         public int CombatTargetTickAccumulator;
         public byte TargetStatusEffectBitmask;
